requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2
django-bootstrap4==1.0.1
django-crispy-forms==1.8.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-registration==3.0.1
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.3
soupsieve==1.9.4
sqlparse==0.3.0
whitenoise==4.1.4
